[pulse] Use the same fields for representing lambda captured variables in Pulse as in the frontend
Summary: In this diff we used the same field names for lambdas in Pulse as they were added in the frontend. We make sure to only change the names of the lambda fields and not the blocks for now. I will also refactor the blocks later and then the names of their fields will be changed too.

Reviewed By: skcho

Differential Revision: D48688335

fbshipit-source-id: b0d706d583fddbd1a0032585edcb94be010ab55d