[textual] make load/store type annotations optionnal
Summary:
We believe we could require less type annotations in Textual, and then
simplify the programmer and frontend work. This first diff introduces the optional type
and adapt the codebase.

If a program contains type holes, transformation to SIL will (probably) fail for now. A next diff will deal with type inference.

Reviewed By: artempyanykh

Differential Revision: D47259579

fbshipit-source-id: f36f524b4526072396cf7fdf14cd3f193a9e633f