[textual][parsing] lexing hack to perform a bit of lookahead with Menhir
Summary:
Lookahead is not always possible in Menhir. Here we add a new parsing rule
 for Textual expressions that exercizes the problem.

Menhir can not figure how to parse an expression like `IDENT DOT ...` because it may be the beginning of a `var.f()` (load and call) expression or a `classname.f()` expression (call).

We solve the problem by using a classic lexing hack:
 a sequence `IDENT (` is directly recognized with a specific token.
 This diff just introduce the hack and makes sure we have neither parsing ambiguities, or
 regressions. A next diff will use the new parsing rule.

Reviewed By: jvillard

Differential Revision:
D48729932

Privacy Context Container: L1208441

fbshipit-source-id: 7c50480476825e4f1a863bd8883d341ec38489bc