[unnecessary copy] Only report on assignment/intermediate for const value types
Summary:
When a copy assignment or intermediate copy is made from a `const &`, the copy is either unavoidable or not straightforward to fix. We already added a check for this specifically for parameters in D44533841. However, we were missing it for local variables.

This diff ensures that we only report on const value types, not const references.

Reviewed By: skcho

Differential Revision: D48866610

fbshipit-source-id: fc40b14f53f000312e48cfd0eb4177a2c3ea1263