[frontend] Revise the translation of list-initialization
Summary:
This diff revises the translation of list-initialization in C++ to support classes that have super
classes.

In the previous implementation, it tried to initialize all fields with the initialization statements
one by one, i.e. by `List.map2`. However, it did not work well when there is a super class, because
the initialization statements for the fields of the super class are under one `InitListExpr`
statement, i.e. one-depth deeper in its AST. This diff applies the initializer statements correctly.

Reviewed By: ezgicicek

Differential Revision: D49099105

fbshipit-source-id: 1fda8e0a734be979f81b345519078b4820e823dc