[ppx] push normalize inside hash_normalize
Summary:
As explained in D49230154 , now the signature generated by the deriver is:

```
    val hash_normalize : t -> t

    val hash_normalize_opt : t option -> t option

    val hash_normalize_list : t list -> t list
```

To help do this, we add a signature deriver for use in `.mli` files.

This will help generating recursive normalizers going forward, as it will avoid the need for recursive functor applications.

Reviewed By: skcho

Differential Revision:
D49265502

Privacy Context Container: L1122176

fbshipit-source-id: 57d641fe5d7a44628b557f2a210c2e7ba63a7121