[infer/py] Add support for qualified name
Summary:
So far, we only supported "short" names for global entities (like
`import foo`). With nested names, modules and such, we have to deal with
qualified names (like `import foo.bar`) or deref sequences (like `a.b.c.f()`)

This diff prepares for that by rewriting a lot of `PyEnv` to index its
map using the newly introduced `Ident.t` type that encodes qualified
names.

A few places still need to be updated, but this is the minimal overhaul
to introduce `Ident.t` and make things behave like before

Reviewed By: davidpichardie

Differential Revision: D49146698

fbshipit-source-id: e6b37daaffa1447893fedac61afef565e122f203