[infer/py] improve Ident.t to remember the global/local status of an id
Summary: To correctly load an `Ident.t`, we need to know if its root was a global or a local variable (because we store them in different maps in PyEnv, because the keys are a bit different). See following diff and the addition of `Ident.fold` that will allow to correctly load qualified names.

Reviewed By: davidpichardie

Differential Revision: D49146694

fbshipit-source-id: f58428a670fb32800cbd630784807a0293c7ba7e