[infer/py] ad-hoc support for `hasattr`
Summary:
`hasattr` is a meta-function provided by Python to check if an entity (object, class, instance, module, ...) has an attribute of some name. There's tons of them in Cinder test suites, in decorators we don't really care about, so for now, we'll just ignore them (and display a warning).

Note that unknown decorators are encoded as `python_call` processing the class object, which is not super nice, but is a good state for now. We might want to support more common decorators as we do `abstractclass`.

Reviewed By: davidpichardie

Differential Revision: D49146689

fbshipit-source-id: 4570249fd5a888c432a4e5e022a50cc0d550ded7