[infer/py] Improved support for accessing imported content
Summary:
When one tries to access definitions / variables from another module, we are lacking a lot of information to do this correctly. For now, we do our best to guess if it's a function / variable. This diff enables us to remove all imports errors from Cinder test suite (coverage 14% to > 20%).

This might not be the best solution, but it enables me to move forward in supporting more of the language. There are on-going discussions with davidpichardie to change the encoding (ideas around encoding modules as proper types and toplevel definitions as member of this type). We'll see once we face real code what is best.

Note: warnings and commented bits of code in the examples will be updated in subsequent diffs

Reviewed By: davidpichardie

Differential Revision: D49146690

fbshipit-source-id: fe2d09605ddcecb63adfc2217dc168fb31d9cb1b