[infer/py][4/N] Starting to migrate opcode entry points to Result.t
Summary:
Stop raising exceptions on errors, report the message and let infer live. For now, we never "die" at all. The last diff in the stack will introduce a call to exit in a more central location.
For this stack, the pattern is to replace all Logging.die FooError "...." with Logging.foo_error "..."; Error ()

Mostly LOAD and STORE family + a few others

Reviewed By: ngorogiannis

Differential Revision: D49309014

fbshipit-source-id: 3bdc6a81f66e82fb46eb5d9a4646818190e60f30