[infer/py][5/N] More opcode entry points migrated
Summary:
Stop raising exceptions on errors, report the message and let infer live. For now, we never "die" at all. The last diff in the stack will introduce a call to exit in a more central location.
For this stack, the pattern is to replace all Logging.die FooError "...." with Logging.foo_error "..."; Error ()

opcodes around FUNCTION/METHOD

Reviewed By: ngorogiannis

Differential Revision: D49309009

fbshipit-source-id: ecb310e4e50a45ee760772ae649f012153bbd110