[infer/py][6/N] Migrating remaining opcode entry points to Result.t
Summary:
Stop raising exceptions on errors, report the message and let infer live. For now, we never "die" at all. The last diff in the stack will introduce a call to exit in a more central location.
For this stack, the pattern is to replace all Logging.die FooError "...." with Logging.foo_error "..."; Error ()

Reviewed By: ngorogiannis

Differential Revision: D49309013

fbshipit-source-id: 26fe24d9e70858669bfd1a0bb5b5be4844af506d