[infer/py][7/N] Migrating most of PyTrans.ml to Result.t
Summary:
Stop raising exceptions on errors, report the message and let infer live. For now, we never "die" at all. The last diff in the stack will introduce a call to exit in a more central location.
For this stack, the pattern is to replace all Logging.die FooError "...." with Logging.foo_error "..."; Error ()

Only `load_cell` & al is remaining, and will require more refactoring
effort

Reviewed By: ngorogiannis

Differential Revision: D49309010

fbshipit-source-id: 78ef9a0a3e10069171ec03bd1fd1e1b944358167