[infer/py][8/n] Migrating the rest of PyTrans.ml to Result.t
Summary:
Stop raising exceptions on errors, report the message and let infer live. For now, we never "die" at all. The last diff in the stack will introduce a call to exit in a more central location.
For this stack, the pattern is to replace all Logging.die FooError "...." with Logging.foo_error "..."; Error ()

Reviewed By: ngorogiannis

Differential Revision: D49309006

fbshipit-source-id: 3360cb215e9ef9a10a7d6de35a76ed2fb1a040e5