[infer/py][9/9] Remove most of the remaining L.die in src/python
Summary:
Stop raising exceptions on errors, report the message and let infer live. For now, we never "die" at all. The last diff in the stack will introduce a call to exit in a more central location.
For this stack, the pattern is to replace all Logging.die FooError "...." with Logging.foo_error "..."; Error ()

A few are left for very edge cases. And we're done !

I updated integration to stop crashing workers in case of error while still reporting the error.

Reviewed By: artempyanykh

Differential Revision: D49309008

fbshipit-source-id: 1381df9f4aba921dfda52cfc61a1366d58aae082