LineageShape: use WTO-based iteration strategy
Summary:
Having switched the Lineage analysis from the RPO to the WTO iteration strategy, we now also make the
switch for LineageShape.

The shape analysis fixpoint-reaching being simpler, especially in the presence of loops, this diff
has no measurable impact on a sample codebase (contrary to Lineage). We hope that it will however
make the code less surprising. We also assume that, if the overhead of computing the WTO isn't
observable, then this strategy will always be better (or not worse) should the domain become more
complex in the future.

Reviewed By: rgrig

Differential Revision: D49508834

fbshipit-source-id: 1a5a790186ac7075cdfaa73b4c71eaa252cb81ec