[infer/py][1/3] Introduce support for function with default arguments
Summary:
Introducing a new kind of cell to keep a list on the stack. This will be used in the next diff to generate specialised versions of functions with default arguments inlined

A known issue is around primitive types vs objects as default arguments. The current way we store them won't work for objects.

Reviewed By: ngorogiannis

Differential Revision: D49494735

fbshipit-source-id: b03fe58b820913db615c002979278b0c1dd6edb4