[uninit][hack] Add semantic model for hhbc_cls_cns
Summary:
This diff adds the semantic model for `hhbc_cls_cns` that reads the class constant field. For example, when a class `A` has a constant `FIELD` and it is read, Pulse's internal state becomes like below:

```
// ret = hhbc_cls_cns( class A, field )

roots={ ret=v9,
        &Uninit::A$static=v7, ... };

mem  ={ v7 -> { FIELD -> v8 },
        v8 -> { * -> v9 } ... };
```

Note that there is the global variable `Uninit::A$static` and its constant value is stored as its field access.

Reviewed By: davidpichardie

Differential Revision: D49727563

fbshipit-source-id: 5335895d30a9e7021cdb0bf04bd2056dd3276676