[textual] Distinguish Prune node kinds
Summary:
Cost analysis needs to distinguish when a node is a Prune node to determine loop bounds. Textual translated all nodes as `Stmt_node` before. This diff tags nodes that contain Prune instructions as `Prune_node`.

Note that AFAIC Textual can't differentiate whether a Prune instruction is for true or false branch (it is always set to true) hence we also don't have that info for `Prune_node` kind.

Reviewed By: davidpichardie

Differential Revision: D49777619

fbshipit-source-id: 3d1d7fa49ce3dda0b7a48f69cf3b1e0bea89b40d