Add a few Pulse erlang comparison tests
Summary:
Exploratory work on Topl showed a limitation/wrong behaviour on some pulse models of erlang
comparisons. Comparing an any-typed value with anything else than an integer (which is precisely
supported) will assume the values to be incompatible, thus for instance always returning `false` for
equality testing.

We add a few tests that exhibit this behaviour on the model for equality. Fixes will come on later
diffs.

Reviewed By: hajduakos

Differential Revision: D49785385

fbshipit-source-id: 82c257b7ef83c766e56acd28163e8fe7cdc6bfa4