Fix Pulse erlang comparisons between {atom, any} and any
Summary:
When comparing two values `x` and `y`, when `x` is an integer and `y` can have any type (usually an
unknown value, such as a function parameter), the Pulse erlang model splits cases on `y` being an
integer or not, and then uses an integer and incompartible comparison on the according cases.

Atoms didn't implement this, but instead of just being considered as unsupported, the code used to
consider that `(x : Atom)` and `(y: Any)` were incompatible. This leads to `x == y` being interpreted
as always `false`, which is neither sound (should be unconstrained) or complete (should be
bottom/unreachable) when `y` can be anything.

We fix this by precisely supporting `Atom` vs. `Any` comparisons the same way we do for integers.

We incidentally fix `Any` vs unsupported types (such as lists) comparisons by having them unsupported
instead of assuming the types are distinct. `Any` vs `Any` is also concerned by this incidental fix.

Remark: this does transform some false negative tests into false positive, which at first sight would
seem to be inconsistent with Pulse goals. This is because equality on unsupported types returns an
unconstrained value instead of stopping the execution, which could be considered as another
change (the reason for this being that stopping the execution would lead to too many false negatives
with the current any-type-splitting logic). Also note that the behaviour, although less precise, is
more consistent: tests that check if the result of an unsupported call is `true` or `false` should
behave the same, which wasn't the case before.

Reviewed By: rgrig

Differential Revision: D49831551

fbshipit-source-id: 67043bd897f62691b1fa1fe0ad029d004560e9c2