[infer/py] more support for SETUP_ANNOTATIONS (toplevel / functions)
Summary:
SETUP_ANNOTATIONS was limited to class declarations, but it can happen anytime some annotation is written (top level, local variables, ...) So we support it eslewhere.

Since it's semantic is to setup a `__annotations__` dict if not present already, we also take care to store them in our state only if SETUP_ANNOTATIONS has been run. Otherwise, this `__annotations__` dictionary might be some user defined one, not under our control

Reviewed By: ngorogiannis

Differential Revision: D49902958

fbshipit-source-id: f592b693af057ae88e76da347cc68ecd077c9fbc