[cost][hack] Add basic support and tests for cost analysis to Hack
Summary:
Adding experimental cost analysis support for Hack. The underlying cost model is the same for now. We might change this to account for only prune nodes in the future.

We make sure to not report costs for hack builtins or models as these might change frequently.

Currently, analysis for loops doesn't work as loop control analysis and inferbo models for comparison are missing.

Reviewed By: davidpichardie

Differential Revision:
D49904439

Privacy Context Container: L1122176

fbshipit-source-id: 2952af10f7b93af10f268e979460821382d3dc90