Add initial revision reference stability checker, only covering a selection of map functions
Summary:
The underlying model for a map is simply

```
v0 -> { __infer_map_pair -> v1 }
v1 -> { first -> v2, second -> v3 }
```

When we call `map.at`, we return `v3`. When we call `map.find`, we return an abstract value pointing to `v1` (iterator). When we call an invalidation function, we invalidate all of `v1`, `v2`, and `v3`, and create new abstract values to represent the map.

As of this diff, `clear`, `rehash`, `reserve`, `at` and `find` should be covered for `F14FastMap` (for `find`, so long as the underlying map type is `F14VectorMap`).

If ever we extend this analysis to maps that only invalidate iterators (not references, `std::unordered_map`, `folly::F14NodeMap`), we could reuse almost everything, and only invalidate `v1`, not `v2` and `v3`.

Reviewed By: skcho

Differential Revision: D49958507

fbshipit-source-id: c4d5644418008f531c32ba6ef68f4c7d76baf5b0