[infer/py] initial support for nested classes
Summary:
Initial attempt at supporting nested classes. This diff is probably incomplete, but the new test case works as expected ;)

A design question is left: should we introduce some mangling in the nested class name (`foo::A` vs `foo::$nested::A`) in case there might be collisions ? I don't think so, but I should make sure.

The main changes of this diff is how we update the `module_name` that is threaded down the line. We now have to add the function names too, so some splitting has been updated (see support for `super()`).

Also, global variables need now a small fix to be set in the right module, when accessed from a function. This is part of this diff too.

Reviewed By: davidpichardie

Differential Revision: D49902970

fbshipit-source-id: 42f972bd39b874ee48d97eece2ad686f61bb738a