[textual}[hack] variadic functions in a trait
Summary:
We have required Textual variadic functions to have the .variadic annot only on their
last parameter, but the way the Hack frontend generates trait functions conflicts with this decision.
A trait function will always be given an extra last parameter `self`.

We identify this situation and relax the verification in this situation.

Reviewed By: vsiles

Differential Revision:
D50076645

Privacy Context Container: L1208441

fbshipit-source-id: 4c47962c9dfaca4ffa4711d20be7c53b7e02f6b5