[taint] Add to the taint attribute that the taint sink may come from a field or pointed to by a value
Summary:
When we add a taint sink attribute to a value in the heap, we also add it to all the values the value points to recursively. However, so far, we were not keeping track of this fact. Now, we are able to express in the tainted attribute that it's a sink because of a predecessor in the heap.

Then, we can differentiate between basic and derived sinks in the next diff to avoid using the value history in the derived case, since this causes false negatives.

Reviewed By: geralt-encore

Differential Revision: D50176052

fbshipit-source-id: 659e8483eefa1e8e68bcbb3e0efd97282ed038e8