[pulse] normalize in pre
Summary:
As the wise rgrig once remarked, it doesn't make much sense to normalise
(abstract) values according to the equality relation in the post conditions but
not the precondition. This remedies the situation along the lines of
rgrig's proposed changes back then, including adding a tree-shape check on
the precondition to detect assumed equalities.

This is part of several changes to make the normalisation more robustly applied
on the abstract state and doing this change first seems to lead to the least
amount of disruption in the tests.

Reviewed By: rgrig

Differential Revision: D49640217

fbshipit-source-id: 75d6230082440c7cf56b6f16b366840e20a6fd9a