[pulse] stitch caller->callee histories using cell ids
Summary:
When applying a summary, pulse stitches together caller histories to the ones
in the callee. But the way we do that is based on the *values* in the callee
alone. However, histories are not attached to values but to memory locations,
eg `x->f->g` (which may have the same value as `z` but for completely different
reasons).

This makes it so we stitch histories together correctly. This is especially
important after having introduced normalisation in the pre (D49640217).

The way it works is to identify each memory location ( = access paths) with a
unique id and use those to know what caller history to apply. Each history
should now come with a "cell id". It is implemented as another variant in
histories that should be at the top, except if there's been a binary operation
(TODO for later). It's a little brittle but should already work in many cases.

This also revises how traces in latent issues are propagated up calling contexts: we want to update  them to include caller history so that, eg, null ptr errors can find invalidation events that happened in the context. In exchange, we remove the "calling context" part of the trace which has duplicated info.

Reviewed By: skcho

Differential Revision: D49868951

fbshipit-source-id: b73ccae86910eb67e21cabe20af9511f2cb259f1