[pulse][interproc] add attributes separately and correctly
Summary:
Attributes from callee summaries were added as part of adding arithmetic
constraints in the precondition, and as part of adding memory edges in the
post. This is messy and neither is correct as there could be more attributes
that weren't added (in the post this was corrected by adding all "remaining
attributes" in an additional phase).

Instead, enumerate attributes from the pre/post in their own phases when
applying summaries.

Reviewed By: dulmarod

Differential Revision: D50087874

fbshipit-source-id: 11823e462b0fc5aaaeebf04cab8975e70c0b2cef