[pulse][interproc] simpler and faster way to apply post
Summary:
Applying the post-condition of a callee summary followed a similar pattern as
materialising the precondition:
- apply starting from the actuals
- apply starting from the globals
- etc.
- apply starting from remaining pre addresses (slightly weird)

This isn't needed: we can just apply the post condition of each address in the
post map one by one. To get sensible histories we start with the addresses in
the precondition, then apply the "new" addresses that appear only in the post.

Besides simplifying the code, this gets rid of a potentially-costly computation
of reachable addresses in the precondition (which wasn't even needed anyway
since the set of reachable values is already known at this point as the domain
of [subst]).

Reviewed By: skcho

Differential Revision: D50087872

fbshipit-source-id: ac13ec2455e959d867d8563fd9300ebf8cbb19f0