[pulse][interproc] delay visiting array accesses in summaries until after arithmetic
Summary:
When applying a callee's precondition we need to match array indices against
array access edges in the caller. These are more sensitive to arithmetic
reasoning than other values so at the moment a costly (and complicated)
normalization phase is done to canonicalize these accesses.

In this diff, we delay exploring array accesses until
1. the rest of the pre has been explored
2. the arithmetic conditions of the precondition have been applied

This avoids the need for normalisation and in fact seems to work better.

Reviewed By: dulmarod

Differential Revision: D50087870

fbshipit-source-id: 41f3b64b03e95a39c177b9240d292316dec33809