Pulse erlang: assume same types on comparisons vs any
Summary:
The recent fix on comparisons of atom vs any-typed erlang values led to the any-typed value being
case-split on being an atom or not, then the atom or incompatible comparator was used on each case.

This is correct but happens to generate many disjunctions, which reduces a lot the coverage of the
erlang analysis on real-sized codebases.

With this diff, we only keep the atom case (and do the same for the other supported type,
Integer). This means we assume that the any-typed value is an atom and them use the atom
comparison. This is a stricter under-approximation at the comparison step, but reducing the following
amount of disjunctions actually makes the analysis more covering.

Reviewed By: mmarescotti

Differential Revision: D50299735

fbshipit-source-id: 2a107ab9cf57a8466ae82551098e8f544313d435