[infer/py][IR/CFG] Add support for SETUP_WITH and minimal exception handling
Summary:
Last remaining opcodes we support in PyTrans, related to `with ... as ...:` and very simple exception (`throw`) support.

While testing with `return/raise` in loops, I realized that the Python compiler often insert "dead" jump just after a terminator. Unfortunately, some stack managment done just before (e.g. a POP_TOP before a  RETURN) is messing with the stack size, and this introduces back-edges that have a stack not compatible with their target labels. To fix I propose the following changes in this diff too:
- before a node is processed, erase the stack content, as it should have been captured by the next label we encounter
- after a node is processed, skip all instructions that are not "labeled"

Reviewed By: ngorogiannis

Differential Revision: D50223087

fbshipit-source-id: ef084844f0f9f10556148e7540d192001b19d6ce