[specialization] allowing partial specialization
Summary:
when a callee needs specialization we scan the caller state and check if we have all what is needed in our hands.
If not, we run a specialized analysis of the callee the partial information we have, while before we were keeping the default
summary and wait for a full specialization.

Drawback: more specialized summary can be computed for a same callee.
Our first experiments shows a x3 efficiency penalty on some targets.

We prefer to gate this feature with a special option.

Reviewed By: skcho

Differential Revision:
D49633651

Privacy Context Container: L1208441

fbshipit-source-id: 657167c7486f4e201e8ecec4b043080006caf07a